Fix docstrings for __future__ imports
By splitting the (default header + copyright notice) and (some other information) strings at the top of a file across two multi-line strings, only the first of those strings is classified as the file docstring. Consequently, if a "from __future__" import call is placed after the second string, a SyntaxError may be raised, due to the narrow set of items that are permissible to appear before such a statement.